requirements.txt
pandas
pydantic>=2.6.0
python-dotenv
requests
tqdm
sqlglot
